[Private Network Access] Allow secure same-origin requests
Metrics in dev show that allowing same-origin requests might eliminate
~50% of the ignored preflights, so implementing this behind a flag and
adding a virtual test suite. We already have same-origin tests which are
mostly "treat-as-public to local". Their PreflightBehavior is updated to
optionalSuccess so that the test server doesn't fail the final request
due to lack of preflights.

This CL also splits out the treat-as-public test cases from some tests so that they can have their own expected outputs.

Bug: 1382068
Change-Id: Idb8cbb6b5470565e97f844ff54eea9b534b3f0ce