[Multi Apps] WPT cleanup
Removes the use of "document.location.origin" outside of tests that
check for the use of full urls. Renames the parameter "sub-app-id" and
"/sub/app/id" to "/sub-app" to be consistent across the tests and call
the API with relative urls. I also put any literal dictionary keys in
quotes to make it more obvious where variables are used (rather than
just identifiers in a dictionary).

Bug:b/268613461
Change-Id: Idf6d236af214c896dd9ed949806ac44b41c3d550