view-transition: Handle display:none on documentElement.
display:none on the documentElement causes it to not generate a
LayoutObject. The current code incorrectly assumes this can never
happen. This can happen during exit/entry navigations.

For cases where documentElement->GetLayoutObject disappears during
a transition, we detect that by checking if we have any target_elements
in the list.

R=bokan@chromium.org, khushalsagar@chromium.org

Fixed:1410303
Change-Id: I1e065877088fe368e4aa717b816fb08c70560029