vt: iframe scrollbars paint last during transition
In a main frame, scrollbars paint during the kOverlayOverflowControls
phase while in an iframe they paint during kSelfBlockBackgroundOnly. The
former causes the scrollbars to avoid being part of the root snapshot
due to the way we handled https://crbug.com/1303081. The latter causes
the scrollbar to be painted into the iframe's root snapshot which is
unintended.

This CL changes the paint order of iframe scrollbars to match the main
frame for the duration of the transition so that snapshots behave the
same way.

Bug: 1407299
Change-Id: Ida74507e59b16f64134807ed56487fc6fb569e1c